warning fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@15176 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Toaster/src/toast_notification.cpp
@@ -56,7 +56,7 @@ HRESULT ToastNotification::CreateXml(_Outptr_ ABI::Windows::Data::Xml::Dom::IXml
 
 	xmlDestroyNode(xmlToast);
 
-	CHECKHR(xmlDocument->LoadXml(StringReferenceWrapper(xtmp, xlen).Get()));
+	CHECKHR(xmlDocument->LoadXml(StringReferenceWrapper(xtmp, (UINT32)xlen).Get()));
 
 	return xmlDocument.CopyTo(xml);
 }
